news-loader/requirements.txt
requests==2.28.1
elasticsearch==8.7.0
google-cloud-bigquery
pandas
db-dtypes